+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
   *Daniel Schierbeck*
 
+* Include the superclass' dependencies in a presenter's dependency list.
+
+  *Daniel Schierbeck*
+
 ### Curly 0.9.0 (June 4, 2013)
 
 * Allow running setup code before rendering a Curly view. Simply add a `#setup!`